# docs/swagger.yaml
basePath: /api
definitions:
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse:
    description: Структура сообщения об ошибке
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  ? github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.OkResponse-github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto_MeResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.MeResponse'
      status:
        type: string
    type: object
...
  ? github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.OkResponse-github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto_UserInfoResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.UserInfoResponse'
      status:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.LogInRequest:
    description: Данные для входа в систему
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: p@ssword123
        type: string
    type: object
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.MeResponse:
    description: Информация об авторизованном пользователе
    properties:
      profile:
        $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.ProfileResponse'
      user_link:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.ProfileResponse:
    description: Полная информация о профиле пользователя
    properties:
      avatar_url:
        example: https://storage.yoursite.com/avatars/123.jpg
        type: string
      description_user:
        example: Люблю писать код на Go и проектировать микросервисы
        type: string
      display_name:
        example: Ivan Ivanov
        type: string
      email:
        example: ivan.ivanov@example.com
        type: string
      link:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.RegisterRequest:
    description: Модель регистрации нового пользователя
    properties:
      display_name:
        example: Ivan Ivanov
        type: string
      email:
        example: ivan@mail.com
        type: string
      password:
        example: securePassword
        type: string
      repeated_password:
        example: securePassword
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.UserInfoResponse:
    description: Информация о пользователе после авторизации
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      display_name:
        example: Ivan Ivanov
        type: string
      email:
        example: ivan@mail.com
        type: string
      link:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
host: clac-clac.ru
info:
  contact: {}
  description: Наше API
  title: NeXuS
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 'Аутентифицирует пользователя по email и паролю. При успехе устанавливает
        session_id cookie. Rate limit: 5 попыток в минуту.'
      parameters:
      - description: Email и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.LogInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход, cookie session_id установлен
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.OkResponse-github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto_UserInfoResponse'
        "400":
          description: Некорректный формат запроса, email или пароль
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "404":
          description: Пользователь не найден или неверные учётные данные
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "429":
          description: Слишком много попыток входа
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      summary: Вход (Login)
      tags:
      - Auth
...
  /me:
    get:
      description: Возвращает 200, userLink и профиль если сессия активна (пользователь
        авторизован). Используется для проверки состояния авторизации на клиенте.
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь авторизован
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.OkResponse-github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto_MeResponse'
        "401":
          description: Сессия отсутствует или истекла
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при получении профиля
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      security:
      - sessionCookie: []
      summary: Проверка авторизации (Me)
      tags:
      - Auth
  /oauth/vk:
    get:
      description: Обменивает временный code от VK на access_token, находит или создаёт
        пользователя, создаёт сессию. Всегда отвечает HTTP 302 редиректом на фронтенд
        — при успехе и при ошибке.
      parameters:
      - description: Временный OAuth-код от VK
        in: query
        name: code
        required: true
        type: string
      responses:
        "302":
          description: 'Редирект с ошибкой: ?code=400 (code пустой), ?code=502 (VK
            недоступен), ?code=500 (ошибка обработки или создания сессии)'
      summary: VK OAuth Коллбэк
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: 'Создаёт новый аккаунт. Пароли должны совпадать, email должен быть
        уникальным. При успехе устанавливает session_id cookie. Rate limit: 5 попыток
        в час.'
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Аккаунт создан, cookie session_id установлен
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.OkResponse-github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto_UserInfoResponse'
        "400":
          description: Некорректный формат запроса, email/пароль; пароли не совпадают
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "404":
          description: Пользователь не найден (ошибка при создании сессии)
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "429":
          description: Слишком много попыток регистрации
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      summary: Регистрация
      tags:
      - Auth
...
swagger: "2.0"